refactor(flagship): move flag IaC out of db/resources.ts into the feature folder (#857)
`apps/web/worker/db/resources.ts` co-owned two unrelated alchemy `bind()`-ables:
the `PhoenixDb` D1 store and the entire Flagship flag-IaC surface. The module
name asserts "owns the database store" but held a flag-IaC surface too, and
`features/flagship/Flagship.ts` reached back across the seam to import the
`Flagship` app declaration from `db/` — a feature module depending on `db/`
purely for a flag-IaC symbol (an inverted dependency direction).

Relocate the `Flagship` app + flag declarations (`demoTargetingFlag` +
`DEMO_TARGETING_*`, `panoDraftSaveFlag`, `PANO_DRAFT_SAVE_FLAG`, the
`PANO_DRAFT_SAVE` re-export) into `features/flagship/resources.ts`, beside their
evaluator. `db/resources.ts` now owns only `PhoenixDb`. Import-rewrite only — no
behavior change: the same Flagship app + flags are provisioned with identical
defaults/targeting. The flag key stays in `src/flags/keys.ts` (out of scope).

Updated the flag-IaC pointers in `.patterns/feature-flags*.md` and
`src/flags/keys.ts` to the new home.

Fixes #857
