+++ apps/web/src/flags/keys.ts
@@ -1,6 +1,7 @@
 /**
  * Flag-key constants shared by both halves of the codec — the client (`FlagGate` /
- * `useFlag`) and the server (the IaC declaration in `worker/db/resources.ts` + the
+ * `useFlag`) and the server (the IaC declaration in
+ * `worker/features/flagship/resources.ts` + the
  * mutation gate). A plain-string module (no alchemy/React import) so it is safe in
  * the worker bundle AND the SPA bundle, mirroring `src/lib/mutationErrorCodes.ts`.
  * One home per key means the gate and the declaration can never name different
