+++ apps/web/worker/features/pano/mutations.ts
@@ -15,10 +15,10 @@
 import {CurrentUser, Fate, Unauthorized} from "@kampus/fate-effect";
 import {Effect} from "effect";
 import * as Schema from "effect/Schema";
-import {PANO_DRAFT_SAVE} from "../../db/resources.ts";
 import {WorkerLivePublisher} from "../fate-live/protocol.ts";
 import {Flags} from "../flagship/Flags.ts";
 import {FlagsContext, makeRequestFlagsContext} from "../flagship/FlagsContext.ts";
+import {PANO_DRAFT_SAVE} from "../flagship/resources.ts";
 import {Bookmark} from "./Bookmark.ts";
 import {
 	CommentNotFound,
